Small test of subcircuit extraction and evaluation passing
