+++ tests/simulator/partial_evaluation.cpp
@@ -7,6 +7,7 @@
 
 #include "coreir/simulator/output.h"
 #include "coreir/simulator/simulator.h"
+#include "coreir/simulator/subcircuit.h"
 #include "coreir/simulator/utils.h"
 #include "coreir/simulator/wiring_utils.h"
 
@@ -160,166 +161,6 @@ namespace CoreIR {
       }
     }
   };
-
-  bool foldConstants(CoreIR::Module* const mod) {
-    if (!mod->hasDef()) {
-      return false;
-    }
-
-    ModuleDef* def = mod->getDef();
-    Context* c = mod->getContext();
-
-    auto driverMap = signalDriverMap(def);
-    auto receiverMap = signalReceiverMap(def);
-
-    for (auto instR : def->getInstances()) {
-      if (getQualifiedOpName(*(instR.second)) == "coreir.const") {
-        Instance* inst = instR.second;
-        cout << "Found constant to fold = " << inst->toString() << endl;
-
-        vector<Select*> receivers =
-          getReceiverSelects(inst);
-
-        cout << "Connections" << endl;
-        for (auto sel : receivers) {
-          cout << "\tConnects to " << sel->toString() << endl;
-        }
-      } else if (getQualifiedOpName(*(instR.second)) == "coreir.mux") {
-        Instance* inst = instR.second;
-
-        cout << "Found mux " << inst->toString() << endl;
-        auto wbs = inst->sel("sel")->getConnectedWireables();
-
-        assert(wbs.size() == 1);
-
-        Wireable* ptr = *std::begin(wbs);
-
-        cout << "Conneted to " << ptr->toString() << endl;
-
-        assert(isa<Select>(ptr));
-
-        Wireable* src = extractSource(cast<Select>(ptr));
-
-        if (isa<Instance>(src) &&
-            (getQualifiedOpName(*(cast<Instance>(src))) == "coreir.const")) {
-
-          Instance* srcConst = cast<Instance>(src);
-          //cout << "Found constant mux" << endl;
-
-          BitVec val =
-            (srcConst->getModArgs().find("value"))->second->get<BitVec>();
-
-          //cout << "value = " << val << endl;
-
-          Select* bitSelect = cast<Select>(ptr);
-
-          string selStr = bitSelect->getSelStr();
-          Wireable* parent = cast<Select>(bitSelect->getParent())->getParent();
-
-          // cout << "Parent = " << parent->toString() << endl;
-          // cout << "Src    = " << src->toString() << endl;
-          assert(parent == src);
-          assert(isNumber(selStr));
-
-          int offset = stoi(selStr);
-
-          uint8_t bit = val.get(offset);
-
-          assert((bit == 0) || (bit == 1));
-
-          Select* replacement = nullptr;
-          Select* toReplace = inst->sel("out");
-          if (bit == 0) {
-            replacement = inst->sel("in0");
-          } else {
-            assert(bit == 1);
-            replacement = inst->sel("in1");
-          }
-
-          //cout << "Receivers of mux output to rewire" << endl;
-          for (auto sel : drivenBy(toReplace, receiverMap)) {
-            //cout << "\t" << "sel = " << sel->toString() << endl;
-
-            auto target = driverMap[sel];
-
-            //cout << "\tsel driver = " << target->toString() << endl;
-
-            Select* val =
-              cast<Select>(replaceSelect(toReplace,
-                                         replacement,
-                                         cast<Select>(target)));
-
-            //cout << "replacement select = " << val->toString() << endl;
-
-            auto driver = map_find(cast<Wireable>(val), driverMap);
-            // Select* driver = nullptr;
-            assert(driver != nullptr);
-
-            //cout << "replacement select driven by " << driver->toString() << endl;
-
-            //cout << "connecting " << sel->toString() << " <--> " << driver->toString() << endl;
-            def->connect(sel, driver);
-          }
-
-          assert(replacement != nullptr);
-
-          def->removeInstance(inst);
-
-        }
-            
-      }
-    }
-    return true;
-  }
-  
-  void registersToConstants(CoreIR::Module* const mod,
-                            std::unordered_map<std::string, BitVec>& regValues) {
-    if (!mod->hasDef()) {
-      return;
-    }
-
-    ModuleDef* def = mod->getDef();
-    Context* c = mod->getContext();
-    for (auto instR : def->getInstances()) {
-      auto inst = instR.second;
-      if (getQualifiedOpName(*inst) == "coreir.reg") {
-
-        cout << "Found register = " << inst->toString() << endl;
-
-        if (contains_key(inst->toString(), regValues)) {
-
-          BitVec value = regValues.find(inst->toString())->second;
-
-          cout << "Replacing register = " << inst->toString() << endl;
-          cout << "Connected wireables = " << endl;
-          for (auto wb : inst->getConnectedWireables()) {
-            cout << "\t" << wb->toString() << endl;
-          }
-
-          string cName = inst->toString() + "_const_value";
-          Instance* constR =
-            def->addInstance(cName,
-                             "coreir.const",
-                             {{"width", Const::make(c, value.bitLength())}},
-                             {{"value", Const::make(c, value)}});
-
-          Select* regOutSel = cast<Select>(inst->sel("out"));
-          Select* constOutSel = cast<Select>(constR->sel("out"));
-
-          for (auto& conn : def->getConnections()) {
-            Wireable* connFst = replaceSelect(regOutSel, constOutSel, conn.first);
-            Wireable* connSnd = replaceSelect(regOutSel, constOutSel, conn.second);
-
-            def->disconnect(conn.first, conn.second);
-            def->connect(connFst, connSnd);
-            
-          }
-
-          def->removeInstance(inst);
-        }
-      }
-    }
-  }
 
   TEST_CASE("Partial evaluation") {
     Context* c = newContext();
